backend/pyproject.toml
[project]
name = "foresttalk-backend"
version = "0.1.0"
description = "Backend API for the ForestTalk MVP."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "deepforest==2.1.0",
  "fastapi>=0.115.5,<1.0.0",
  "pillow>=10.4.0,<13.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "uvicorn[standard]>=0.32.1,<1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
